Add exact results for trig fcns on π
The argument in favor here is that since Julia  has  already gone to the
trouble to encode π exactly, and trig functions are defined in terms of
π, they should give exact answers for this particular constant. `log(ℯ)`
has a similar definition and test that `log(ℯ) === 1`. I decided here
not to require a specific return type as long as it was numerically
exact. (Arguably, we could also have chosen `log(::typeof(ℯ)) = true`.)
I chose `Int` because there is no obvious reason to impose a particular
precision.